adding tests for pysal_utils
